+++ assets/src/scripts/monitoring-location/store/hydrograph-data.js
@@ -68,7 +68,7 @@ const addMedianStatisticsData = function(statsData) {
 };
 
 /*
- * Synchronous Redux action which sets the median data
+ * Synchronous Redux action which sets the groundwater levels data
  * @param {Object} gwLevels
  *      @prop {Object} parameter
  *      @prop {Array} values
@@ -97,7 +97,7 @@ const retrieveIVData = function(siteno, dataKind, {parameterCode, period, startT
         const isCalculatedTemperatureCode = isCalculatedTemperature(parameterCode);
 
         return fetchTimeSeries({
-            sites: [siteno],
+            siteno: siteno,
             parameterCode: getParameterToFetch(parameterCode),
             period: period,
             startTime: startTime,
@@ -194,7 +194,7 @@ export const retrieveMedianStatistics = function(siteno, parameterCode) {
         } else {
             const isCalculatedParameterCode = isCalculatedTemperature(parameterCode);
             const parameterToFetch = getParameterToFetch(parameterCode);
-            return fetchSiteStatistics({siteno: siteno, statType: 'median', params: [parameterToFetch]})
+            return fetchSiteStatistics({siteno: siteno, statType: 'median', parameterCode: parameterToFetch})
                 .then(stats => {
                     let resultStats = {};
                     if (parameterToFetch in stats) {
@@ -226,9 +226,9 @@ export const retrieveMedianStatistics = function(siteno, parameterCode) {
  * @param {String} endTie - ISO 8601 time string
  * @return {Function} that returns a Promise
  */
-const retrieveGroundwaterLevels = function(site, {parameterCode, period, startTime, endTime}) {
+const retrieveGroundwaterLevels = function(siteno, {parameterCode, period, startTime, endTime}) {
     return function(dispatch) {
-        return fetchGroundwaterLevels({site, parameterCode, period, startTime, endTime})
+        return fetchGroundwaterLevels({siteno, parameterCode, period, startTime, endTime})
             .then(data => {
                 if (data.value && data.value.timeSeries && data.value.timeSeries.length) {
                     let values;