+++ assets/src/scripts/monitoring-location/store/hydrograph-parameters.js
@@ -29,7 +29,7 @@ export const retrieveHydrographParameters = function(siteno) {
     return function(dispatch) {
         let fetchPromises = [];
         if (config.ivPeriodOfRecord) {
-            const fetchIVParameters = fetchTimeSeries({sites: [siteno]})
+            const fetchIVParameters = fetchTimeSeries({siteno: siteno})
                 .then(series => {
                     if (series.value && series.value.timeSeries) {
                         const allParameterCodes = series.value.timeSeries.map(ts => ts.variable.variableCode[0].value);
@@ -66,7 +66,7 @@ export const retrieveHydrographParameters = function(siteno) {
             fetchPromises.push(fetchIVParameters);
         }
         if (config.gwPeriodOfRecord) {
-            const fetchGWLevelParameters = fetchGroundwaterLevels({site: siteno})
+            const fetchGWLevelParameters = fetchGroundwaterLevels({siteno: siteno})
                 .then(series => {
                     if (series.value && series.value.timeSeries) {
                         return series.value.timeSeries.reduce((varsByPCode, ts) => {