+++ assets/src/scripts/web-services/instantaneous-values.js
@@ -25,21 +25,28 @@ function getNumberOfDays(period) {
 * @param {String} siteno - the unique identifier for the monitoring location
 * @param {Boolean} isExpanded - if the meta data has additional information
 */
-export const getServiceURLMetaData = function({siteno,  isExpanded}) {
+export const getSiteMetaDataServiceURL = function({siteno,  isExpanded}) {
     return `${config.SERVICE_ROOT}/site/?format=rdb&sites=${siteno}${isExpanded ? '&siteOutput=expanded' : ''}&siteStatus=all`;
 };
 
 /*
 * Get a URL formatted to download data from waterservices.usgs.gov
 * @param  {Array}  sites  Array of site IDs to retrieve.
-* @param  {String}  parameterCode - USGS five digit parameter code
+* @param  {String} parameterCode - USGS five digit parameter code
 * @param {String} period - ISO 8601 Duration
 * @param {String} startTime - ISO 8601 time
 * @param {String} endTime - ISO 8601 time
 * @param {String} format - the data format returned from waterservices.usgs.gov
 * @return {String} The URL used to contact waterservices.usgs.gov
  */
-export const getServiceURL = function({siteno, parameterCode= null, period=null, startTime=null, endTime=null, format=null}) {
+export const getIVServiceURL = function({
+                                             siteno,
+                                             parameterCode = null,
+                                             period = null,
+                                             startTime = null,
+                                             endTime = null,
+                                             format = null
+                                         }) {
     let timeParams;
     let serviceRoot;
 
@@ -64,16 +71,16 @@ export const getServiceURL = function({siteno, parameterCode= null, period=null,
 
 /**
  * Get a given time series dataset from Water Services.
- * @param  {Array}  sites  Array of site IDs to retrieve.
- * @param  {String}  parameterCodes
+ * @param  {String}  sites  Array of site IDs to retrieve.
+ * @param  {String} parameterCode
  * @param {String} period - ISO 8601 Duration
  * @param {String} startTime - ISO 8601 time
  * @param {String} endTime - ISO 8601 time
  * @return {Promise} resolves to an array of time series model object, rejects to an error
  */
-export const fetchTimeSeries = function({sites, parameterCode= null, period=null, startTime=null, endTime=null}) {
-    return get(getServiceURL({
-            siteno: sites,
+export const fetchTimeSeries = function({siteno, parameterCode=null, period=null, startTime=null, endTime=null}) {
+    return get(getIVServiceURL({
+            siteno: siteno,
             parameterCode: parameterCode,
             period:  period,
             startTime: startTime,